[android] INTERNET permission + minor media player fixes
So if this sample is meant to make web requests, it doesn't have the Android permission to do so!

This probably works fine in Debug mode, because we add the permission for you. Debugging requires it.

I also cleaned up a bit of the code setting up the media player. It seems like we were just calling into Java over and over, and I could simplify it.